8/3 TicTacToe Project - Update
I can go ahead and use this version to present my work.  Added api input
forms in place of the button calls I had created.  Still not linked properly
but works with the counter added and no need for ninth click for a tie
alert.  I just set turns to ===8 not ===9.  Work on api and if time try
to use images instead of X's and O's.  Need a break.  SPent so much time
messing with code and just had it nested in wrong place.  Original logic
worked and was as easy as I thought - see temp file in this repo.
s