August 3,2017 - GaProject TicTactoe - to submit
This is the last commit prior to the presentation of my project.  The README.md
file explains where I am at.  The game works.  I will show the various
stages of the games development and explain what still needs to be done
to finish the api portion of the project.