8/14 Ga TicTacToe resubmit
Tried at end of class to redeploy after changing container id and form
ids to be different.  Worked on local host but messed up styling on deployed
version.  This was an attempt to see if api was not working because logic
was assigned to wrong place on html.  Also, tried to get rid of image
placeholders and it just messed up alignment.  Currently GitHub page and
file tree are a mess need to clean them up - probably will help.